// src/types/ContactType.ts
export type ContactType = {
    address: string | null
    avatar: string | null
    company: string | null
    createdAt: number
    email: string | null | null
    first_name: string | null
    gender: string | null
    id: number | null
    last_name: string | null
    note: string | null
    phone: string | null
    telegram: string | null
    updatedAt: number
}